Rename view directory to ui
