+++ app/src/main/java/com/senex/androidlab1/ui/fragments/list/recycler/ListRecyclerAdapter.kt
@@ -1,4 +1,4 @@
-package com.senex.androidlab1.views.fragments.list.recycler
+package com.senex.androidlab1.ui.fragments.list.recycler
 
 import android.view.LayoutInflater
 import android.view.View
