+++ app/src/main/java/com/senex/androidlab1/ui/fragments/list/recycler/NoteDiffCallback.kt
@@ -1,4 +1,4 @@
-package com.senex.androidlab1.views.fragments.list.recycler
+package com.senex.androidlab1.ui.fragments.list.recycler
 
 import androidx.recyclerview.widget.DiffUtil
 import com.senex.androidlab1.models.Note
