+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".views.activities.main.MainActivity">
+    tools:context=".ui.activities.main.MainActivity">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/main_toolbar"
